Overhaul of the innards of the parsing algorithm
Remove reference to System.Runtime.CompilerServices.  This was going to be
used so that methods could be decorated with AggressiveInlining attribute,
but alas, it seems one cannot do that in visual studio 2010

Create compile time constant NumberStyles for signed floating point
parsing, which only allows for decimal points and a leading sign.  This
differs from the traditional Float style in that it does not allow leading
and trailing whitespace, two things that could signal an error in parsing.

Remove parameter from many of the private functions.  No need when we have
it as a private variable.

Add ReadInsideBrackets, which parses the innards of the curly bracketes.