ReadInt32 parses from an intermediate string
The handroll method can not handle whitespace and other tokens that could
be in the way.  For instance, what is the defined behavior when x = { 1 }.
Currently the intermediate string ignores everything until the "1", which
is desired.

Add more of an advanced test for buildings from hoi3.

Add special formatting tests for intermediate whitespace.