Exceptions for ReadBracketContent
If the first token read is not a left curly, this does not guarantee that
the bracket content is upcoming, so throw an exception.

Also once the action has been invoked, do a comparison witht he indent
which the bracket content started with.  IF the currentIndent is less than
the startingIdnent, then we know for sure that the action read through the
right curly, so throw an exception.

If the currentIndent and the startingIndent are equivalent, while it is
possible that the action could have overread, we'll assume good faith