Refactor parse to use peekToken
Instead of reading the string and testing for null, use peektoken.

This simplifies the logic somewhat.  Using the peekToken and the ckeck for
the untyped lexer will guarantee that ReadString returns an interesting
result and won't return lexer token in string form.  Note: this is not a
bug fix as if ReadString encountered a token, it would just recursively
call itself.