Convert stringbuilder to manual char array
A 50MB file took approximately 0.85 seconds to parse.  When run through
the VS profiler, the majority of the time was used to call ToString() and
appending casted bytes as characters to the stringbuilder.

The alternative is to manually manage a character array and create a
string from this array.  The same file was profiled with the new way, and
it consistently takes 0.65 seconds to parse, which is a 25% improvement.
The parser did not spend a majority of the time in any one particular
place.  The plurality of the time was spent in the constructor of the
string; however, that was 14.2% of the time time and so I feel that it is
well optimized.