Fix read bracket content bug
In doWhileBracket, the function would exit too early.  It would exit
whenever a right curly bracket was found.  Replaced the complicated
conditional with a more convential loop.

Got rid of the necessary paragraph of a comment preceding the old
conditional

Regression tests added so the same bug doesn't occur again.